Fix gen cpp Resource issue causing compressed_size to be 0
In the generated mappings.cxx, the decltype of the char[] constant is
zero because the type is not realized in the implementation of
mappings.cxx, only in the specific resource declaration.

This can be fixed using a runtime mutation of the mappings file, but in
the meantime the easiest solution is just to use the value calculated by
the Counter in the Resource compressor.

This feature was added and tested with an ugly comment in the generated
source to remind me to go back and change it later.